requirements.txt
python-dotenv==1.0.1
langchain==0.3.13
langchain-community==0.3.13
openfga-sdk==0.9.0
faiss-cpu==1.9.0.post1
langchain-openai==0.3.0
fastapi[standard]>=0.112.0,<0.113.0
uvicorn==0.22.0
Pillow==11.1.0 
lxml==5.3.0
PyPDF2==3.0.1